requirements.txt

plotly
ffastf1
myst_nb
panel